+++ src/include/CMakeLists.txt
@@ -106,21 +106,6 @@ install (FILES "${CMAKE_CURRENT_BINARY_DIR}/kdbconfig.h"
 	       "${CMAKE_CURRENT_BINARY_DIR}/kdbversion.h"
 	 DESTINATION include/${TARGET_INCLUDE_FOLDER})
 
-check_binding_included ("io_uv" IO_UV_INCLUDED SUBDIRECTORY "io/uv" SILENT)
-if (IO_UV_INCLUDED)
-	install (FILES kdbio_uv.h DESTINATION include/${TARGET_INCLUDE_FOLDER})
-endif ()
-
-check_binding_included ("io_glib" IO_GLIB_INCLUDED SUBDIRECTORY "io/glib" SILENT)
-if (IO_GLIB_INCLUDED)
-	install (FILES kdbio_glib.h DESTINATION include/${TARGET_INCLUDE_FOLDER})
-endif ()
-
-check_binding_included ("io_ev" IO_EV_INCLUDED SUBDIRECTORY "io/ev" SILENT)
-if (IO_EV_INCLUDED)
-	install (FILES kdbio_ev.h DESTINATION include/${TARGET_INCLUDE_FOLDER})
-endif ()
-
 add_custom_target (elektra_config_headers
 		   ALL
 		   DEPENDS ${CMAKE_CURRENT_BINARY_DIR}/kdb.h
