+++ src/include/kdbio/glib.h
