+++ src/libs/notification/example/example_notification_async.c
@@ -10,7 +10,7 @@
 #include <kdb.h>
 #include <kdbhelper.h>       // elektraFree
 #include <kdbio.h>	   // I/O binding functions (elektraIo*)
-#include <kdbio_uv.h>	// I/O binding constructor for uv (elektraIoUvNew)
+#include <kdbio/uv.h>	// I/O binding constructor for uv (elektraIoUvNew)
 #include <kdbnotification.h> // notification functions
 
 #include <uv.h> // uv functions
