+++ src/plugins/dbus/dbus.h
@@ -18,7 +18,7 @@
 #include <string.h>
 
 // elektraIoDbus*()
-#include <kdbio_adapter_dbus.h>
+#include <kdbio/adapters/dbus.h>
 
 /**
  * @internal
